# backend/docs/openapi.yml
openapi: 3.1.0
info:
  title: EdTech OS — Multi-Tenant Coaching API
  version: 1.0.0
  description: |
    **Multi-Tenant EdTech OS** — complete REST API for coaching institute management.

    ### Authentication
    All protected routes require a JWT Bearer token obtained from `POST /auth/login`.
    Include it as: `Authorization: Bearer <accessToken>`

    ### Roles
    | Role | Description |
    |------|-------------|
    | `super_admin` | Platform owner — manages all institutes |
    | `coaching_admin` | Institute owner — manages students, teachers, fees |
    | `teacher` | Takes attendance, uploads content, creates live classes |
    | `student` | Views content, fees, live classes |

    ### Tenant Isolation
    Every request is scoped to the authenticated user's `tenantId` from their JWT.
    Never pass `tenantId` in the body — it is always read from the token.

  contact:
    name: EdTech OS Support
    email: support@edtechos.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
  - url: https://api.edtechos.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login

  # ───────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Phone must be 10-15 digits
        errors:
          type: array
          items:
            type: object
...
    TenantListItem:
      type: object
      properties:
        id:              { type: integer, example: 1 }
        name:            { type: string, example: Apex Academy }
        slug:            { type: string, example: apex }
        city:            { type: string, nullable: true, example: Nashik }
        isActive:        { type: boolean, example: true }
        status:          { type: string, example: trial }
        trialEndsAt:     { type: string, format: date-time, nullable: true }
        nextBillingDate: { type: string, format: date-time, nullable: true }
        studentCount:    { type: integer, example: 48 }
    RegisterTenantRequest:
      type: object
      required: [name, slug, adminName, adminPhone, adminPassword]
      properties:
        name:          { type: string, example: Apex Academy }
        slug:          { type: string, example: apex, description: "Lowercase letters, digits, hyphens only" }
        city:          { type: string, example: Nashik }
        contactPhone:  { type: string, example: "919999900001" }
        primaryColor:  { type: string, example: "#2563EB" }
        adminName:     { type: string, example: Rajesh Deshmukh }
        adminPhone:    { type: string, example: "919000000011" }
        adminPassword: { type: string, example: "Coach@123", minLength: 6 }
        plan:          { type: string, enum: [flat, per_student], example: per_student }
        amount:        { type: integer, example: 2500 }
...
    SubscriptionListItem:
      type: object
      properties:
        tenantId:        { type: integer, example: 1 }
        name:            { type: string, example: Apex Academy }
        status:          { type: string, example: active }
        plan:            { type: string, example: per_student }
        amount:          { type: integer, example: 21120 }
        trialEndsAt:     { type: string, format: date-time, nullable: true }
        nextBillingDate: { type: string, format: date-time, nullable: true }

    # ── Lead ─────────────────────────────────────────────────────────────────
...
    CreateSubjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Physics }
    SubjectItem:
      type: object
      properties:
        id:   { type: integer, example: 1 }
        name: { type: string, example: Physics }

    # ── Timetable ────────────────────────────────────────────────────────────
...
    AttendanceRecord:
      type: object
      required: [studentId, status]
      properties:
        studentId: { type: integer, example: 7 }
        status:    { type: string, enum: [present, absent, late], example: present }
    MarkAttendanceRequest:
      type: object
      required: [batchId, date, records]
      properties:
        batchId:     { type: integer, example: 3 }
        timetableId: { type: integer, example: 12 }
        date:        { type: string, format: date, example: "2026-07-09" }
        records:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AttendanceRecord'

    # ── Content ──────────────────────────────────────────────────────────────
    CreateContentRequest:
      type: object
      required: [title, youtubeUrl]
      properties:
        title:      { type: string, example: "Kinematics — Part 1" }
        youtubeUrl: { type: string, format: uri, example: "https://youtu.be/abc123" }
        batchId:    { type: integer, example: 3 }
        subjectId:  { type: integer, example: 1 }
    ContentItem:
      type: object
      properties:
        id:         { type: integer, example: 9 }
        title:      { type: string, example: "Kinematics — Part 1" }
        youtubeUrl: { type: string, example: "https://youtu.be/abc123" }
        batch:      { type: string, nullable: true, example: "JEE 2026 Morning" }
        subject:    { type: string, nullable: true, example: Physics }

    # ── Live Classes ─────────────────────────────────────────────────────────
...
    BroadcastResponse:
      type: object
      properties:
        recipientCount:
          type: integer
          example: 12
          description: Number of users a notification row was inserted for.
    BroadcastToAdminsRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, example: "Platform maintenance tonight" }
        body:  { type: string, example: "The app will be briefly unavailable between 1-2 AM IST for scheduled maintenance." }
        city:
          type: string
          example: Nashik
          description: Omit to broadcast to every coaching_admin platform-wide; set to target only institutes in one city.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: INVALID_TOKEN, message: "Token is invalid or expired" }
    Forbidden:
      description: Insufficient role or plan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            role_forbidden:
              value: { code: FORBIDDEN, message: "Access denied for your role" }
            trial_expired:
              value: { code: TRIAL_EXPIRED, message: "Your free trial has ended. Please subscribe." }
            plan_upgrade:
              value: { code: PLAN_UPGRADE_REQUIRED, message: "This feature requires the Pro plan or higher." }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: NOT_FOUND, message: "Resource not found" }
    ValidationError:
      description: Request body / params failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: VALIDATION_ERROR, message: "phone: Phone must be 10-15 digits" }

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /superadmin/tenants:
    get:
      tags: [Super Admin — Tenants]
      summary: List all institutes
      operationId: listTenants
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all institutes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TenantListItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Super Admin — Tenants]
      summary: Register a new institute
      description: Creates a new tenant with a trial subscription and coaching_admin account in one transaction.
      operationId: registerTenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterTenantRequest' }
      responses:
        '201':
          description: Institute registered
          content:
            application/json:
              example:
                tenant: { id: 3, name: "Mehta Classes", slug: "mehta", city: "Nashik" }
                adminUser: { id: 15, fullName: "Suresh Mehta", phone: "919876543210", role: coaching_admin }
        '409':
          description: Slug already taken
          content:
            application/json:
              example: { code: SLUG_TAKEN, message: "This slug is already in use" }
        '422': { $ref: '#/components/responses/ValidationError' }
  /superadmin/tenants/{id}/suspend:
    patch:
      tags: [Super Admin — Tenants]
      summary: Activate / Suspend an institute
      operationId: suspendTenant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean, example: false }
      responses:
        '200':
          description: Tenant activation state updated
          content:
            application/json:
              example: { id: 1, name: "Apex Academy", isActive: false }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /superadmin/subscriptions:
    get:
      tags: [Super Admin — Subscriptions]
      summary: List all institute subscriptions
      operationId: listSubscriptions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All subscriptions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubscriptionListItem' }
  /superadmin/subscriptions/expiring:
    get:
      tags: [Super Admin — Subscriptions]
      summary: Trials expiring soon
      description: Returns institutes whose trial ends within N days.
      operationId: expiringSoon
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 3, example: 3 }
          description: Look-ahead window in days
      responses:
        '200':
          description: Expiring soon list
          content:
            application/json:
              example:
                - tenantId: 5
                  name: "Newton Classes"
                  status: trial
                  trialEndsAt: "2026-07-12T00:00:00.000Z"

  # ══════════════════════════════════════════════════════════════════════════
  # SUPER ADMIN — LEADS
  # ══════════════════════════════════════════════════════════════════════════
...
  /superadmin/notifications/broadcast:
    post:
      tags: [Super Admin — Notifications]
      summary: Broadcast a notification to coaching_admins
      description: |
        Inserts one notification row per matching coaching_admin, in a single
        query. Omit `city` to reach every institute platform-wide, or set it
        to target only institutes in one city/area.
      operationId: broadcastToAdmins
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BroadcastToAdminsRequest' }
            examples:
              all_admins:
                summary: Broadcast to every institute
                value: { title: "Platform maintenance tonight", body: "Brief downtime 1-2 AM IST." }
              by_city:
                summary: Broadcast filtered by city
                value: { title: "Nashik office closed Monday", city: Nashik }
      responses:
        '201':
          description: Broadcast sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BroadcastResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — DASHBOARD
  # ══════════════════════════════════════════════════════════════════════════
...
  /admin/subjects:
    get:
      tags: [Coaching Admin — Batches & Subjects]
      summary: List subjects
      operationId: listSubjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubjectItem' }
    post:
      tags: [Coaching Admin — Batches & Subjects]
      summary: Create a subject
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubjectRequest' }
      responses:
        '201':
          description: Subject created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubjectItem' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — TIMETABLE
  # ══════════════════════════════════════════════════════════════════════════
...
  /teacher/schedule/today:
    get:
      tags: [Teacher]
      summary: Today's class schedule
      operationId: todaySchedule
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Today's timetable entries
          content:
            application/json:
              example:
                - timetableId: 12
                  batch: "JEE 2026 Morning"
                  subject: Physics
                  startTime: "09:00:00"
                  endTime: "10:30:00"
  /teacher/batches/{batchId}/students:
    get:
      tags: [Teacher]
      summary: Get students in a batch (for attendance)
      operationId: batchStudents
      security:
        - BearerAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema: { type: integer, example: 3 }
      responses:
        '200':
          description: Students in batch
          content:
            application/json:
              example:
                - studentId: 7
                  name: "Aarav Joshi"
                  rollNo: "APX-001"
  /teacher/attendance:
    post:
      tags: [Teacher]
      summary: Mark attendance (bulk)
      description: |
        Saves attendance for all students in a batch for the given date.
        Returns `absentReminders` — pre-built `wa.me` links for absent students'
        parents (only returned if the institute has the `whatsapp_reminders` feature).
      operationId: markAttendance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarkAttendanceRequest' }
            example:
              batchId: 3
              timetableId: 12
              date: "2026-07-09"
              records:
                - { studentId: 7, status: present }
                - { studentId: 8, status: absent }
                - { studentId: 9, status: late }
      responses:
        '200':
          description: Attendance saved
          content:
            application/json:
              example:
                saved: true
                absentReminders:
                  - "https://wa.me/919000000099?text=Aarav%20Joshi%20aaj%20absent%20hai."
  /teacher/content:
    get:
      tags: [Teacher]
      summary: List uploaded video content
      operationId: listContent
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Content list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentItem' }
    post:
      tags: [Teacher]
      summary: Upload a video lecture link
      operationId: createContent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateContentRequest' }
      responses:
        '201':
          description: Content added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
...
  /teacher/doubt-link/{studentId}:
    get:
      tags: [Teacher]
      summary: Get WhatsApp doubt-answer link
      description: |
        **Requires Pro or Elite plan** (`doubt_solving` feature).
        Returns a pre-filled `wa.me` URL from teacher to the student.
      operationId: doubtLink
      security:
        - BearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer, example: 7 }
        - name: text
          in: query
          schema: { type: string, example: "Your doubt about Newton's 3rd law has been answered." }
      responses:
        '200':
          description: WhatsApp link
          content:
            application/json:
              example:
                waUrl: "https://wa.me/919000000013?text=Your%20doubt%20about%20Newton%27s%203rd%20law..."
        '403': { $ref: '#/components/responses/Forbidden' }

  # ══════════════════════════════════════════════════════════════════════════
  # STUDENT
  # ══════════════════════════════════════════════════════════════════════════
...
  /student/videos:
    get:
      tags: [Student]
      summary: List available video lectures
      operationId: listVideos
      security:
        - BearerAuth: []
      parameters:
        - name: subjectId
          in: query
          schema: { type: integer, example: 1 }
          description: Filter by subject
      responses:
        '200':
          description: Video lectures
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentItem' }
  /student/videos/{id}:
    get:
      tags: [Student]
      summary: Get video detail
      operationId: videoDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 9 }
      responses:
        '200':
          description: Video detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentItem' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /student/ask-doubt:
    get:
      tags: [Student]
      summary: Get doubt link to teacher (WhatsApp)
      description: Returns a pre-filled `wa.me` URL from student to teacher.
      operationId: askDoubt
      security:
        - BearerAuth: []
      parameters:
        - name: teacherId
          in: query
          required: true
          schema: { type: integer, example: 4 }
        - name: chapter
          in: query
          schema: { type: string, example: "Newton's Laws of Motion" }
      responses:
        '200':
          description: WhatsApp doubt link
          content:
            application/json:
              example:
                waUrl: "https://wa.me/919000000012?text=Sir%2C%20Newton%27s%20Laws%20mein%20doubt%20hai.%20-%20Aarav%20(Apex%20Academy)"
...